Use HTTP for local storage calls
Storage calls that are handled by the local host should always use HTTP,
and not set up a connecting through stunnel. The SM URL that is passed
to VM.migrate_send in the `dest` parameter, which comes from the return
value of a call to host.migrate_receive, may be an HTTPS URL to the
localhost, which needs to be rewritten upon receiving it.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>