Do not wait for stunnel disconnect
When starting a mirroring process for a disk as part of a storage
migration, xapi establishes the connection to the destination, but hands
it over to tapdisk to do the actual mirroring over NBD. It is crucial
that xapi just hands over the file descriptor and then continues with
other business, without waiting for the connection to finish. This is
how it works for TCP connections now.

When switching to TLS connections, xapi starts an stunnel process as
part of the connection setup, and hands over the stunnel fd to tapdisk.
By default, this functionality then waits for stunnel to finish after
the connection is eventually broken, thus introducing the unwanted
blocking. We fix this by telling stunnel to disconnect, but not wait for
this to actually happen.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>