CA-370578 use subsystemId in NVidia GPU matching
So far, we identified a GPU by its deviceId; this turns out to be
ambiguous and vgpuConfig.xml may contain multiple GPUs with identical
deviceId, causing us to use more vGPU types than we can actually
support.

This patch introduces a new abstract type "gpu" in the VENDOR signature
such that its Nvidia instance can use more information to identify a
GPU.

When matching a GPU found on the PCI bus with GPUs listed in
vgpuConfig.xml we use:

      deviceId match and subsystemId match
  or  deviceId match and subsystemId in vgpuConfig is 0

subsystemId=0 is used by Nvidia as a wildcard and thus subsystemId=0 is
ignored during a match.

The matching for other vendors remains unchanged (and is still based on
deviceId).

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>