CP-40823 Edited Vdi.t in xapi/storage_impl.ml to record vm
The change to the structure of the Vdi module allows for the vm
associated with a dp on a vdi to be recorded. In the calls DP.destroy
and VDI.destroy, a blank quote was given as no vm was available, this
has now been remedied by recording and querying this record so there is a
valid vm when calling SMAPIv3 interfaces.

Signed-off-by: jameshensmancitrix <james.hensman@citrix.com>