CA-370082: Block multiple definitions of certificate-chain in xe cli
User were allowed to define more than one certificate-chain when
installing a host certificate. The implementation picked the first one,
ignoring successive ones.

This could lead to a situation where the host served a certificate with
certificates missing in their chain of trust, making it unverifiable.

Instead detect the situation and block the operation immediately.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>