CA-352073: Ensure all serialized calls can pass rbac checks
Explicitly define a list of tuples with names and values so the rbac
checks can never fail at runtime on differing lengths for these. Instead
now the check will fail at build time.

Calls with defaults allowed the values list to be shorter as they never
bothered to contain the default values. Now the default values are
collected to match all the names of the parameters.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>