CP-41028: Refactor rpc function creation for storage migration
The code that constructs rpc functions has been moved to Storage_utils
and extended with options to turn on certificate checking of the remote
server. The following commits will set those parameters to the right
values depending on whether the storage migration takes place within a
pool, or cross pool.

Also removed the use of the term "local" around this code, as in some
cases these weren't actually used for host-local connections.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>