Date: use conversions to/from Ptime.t, unix_time and formatted strings
Currently the only way to create dates is from floats and strings, since
these do not have a semantic meaning, new names are now used for the
conversions, making it obvious floats represent Unix timestamps and
strings with representation do dates using RFC 3339 and ISO 8601.
Additionally allow conversions from and to Ptime.t, which is actually
the main type that implements datetimes.

There's a new helper function to get the current time in date as a
convenience method.

The origin of Unix time has been renamed to `epoch` from `never`: using
never as a synonym to the beginning of Unix time is not quite accurate
semantically even if the value itself will be exceedingly rare in
practice.

In the future the `{to,of}_float`, `{to,of}_string` and `never` will be
deprecated. Not doing so immediately allows for a smooth transition
path.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>