CA-372128: Avoid extra comparisons in db_cache_types
Previous code tried to avoid creating memory garbage by trying comparing
the old value of a map with the new when updating it.

This unfortunately was done using structural equality instead of
physical (its memory location). This dominated the runtime of xapi under
some conditions.

Instead of trying to optimize use directly the standard library which
does the right thing (physical equality), this reduces CPU cost by half
in systems with thousands of VMs

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>