CA-140252: fix flag handling
The flag was only turning off the debug message but not the code itself.
Indentation now makes it very clear that this is wrong, so lets fix it.

Currently this workaround (from the 6.x era is still on in the .conf
file, but at some point we may want to turn it off)

Signed-off-by: Edwin Török <edvin.torok@citrix.com>