CP-39594: Add datamodel for Gpg_key
This series adds a few functions to manage RPM GPG public keys in dom0.

The 'pool.install_rpmgpgkey' is used to add a new RPM GPG pubkey in a
pool and sync the key to every member host in the pool.
The 'pool.uninstall_rpmgpgkey' is to uninstall a managed RPM GPG pubkey.
The 'pool.sync_rpmgpgkeys' is to sync all managed RPM GPG pubkey
throughout all the hosts in the pool.

With these functions, XAPI users can add a new RPM GPG public key and
then configure it to a repository. In updating of a pool, the RPM GPG
key will be used to verify the metadata of the repository and the RPM
packages in the repository. What's more, the users can uninstall an
existing RPM GPG pubkey and install a new one if the existing on is to
be replaced.

Signed-off-by: Ming Lu <ming.lu@citrix.com>