CP-41028: Store and use verification mode in storage migration state
When starting a storage copy or mirror operation, the call now comes
with a verify_dest parameter, indicating whether the remote server's
certificate must be verified when connecting to it. These calls return
an ID for the operation, which is performed asynchronously.

Any following calls regarding the same operation, e.g. to stop or
cancel, only take this ID as an argument, but may require new
connections to the remote. For this reason, a state record with
connection details is maintained for every operation. This commit adds
the verify_dest parameter to this state, and to all the places in calls
that use the state to make connections.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>