CP-41028: Use Stunnel_client in xapi, not in xenopsd
...to decide whether to use certificate checking or not. This module
uses a local ref to determine whether checking is enabled on the host.
In xapi, this ref is set from a file on startup, and after that it is
updated if checking is enabled/disabled through the API.

Xenopsd currently does not initialise the ref, so checking is always off.
We could initialise the ref on startup, like in xapi, but then xenopsd
still won't get updates if the host-level switch is flipped. Instead,
let xapi worry about the host-level state, by querying Stunnel_client,
and tell xenopsd exactly what to do.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>