CA-342973: use systemd_notify for xapi finishing initialisation
While this does not avoid external clients from sending requests too
early, it allows systemd to know when the service is up and continue
with the boot.

Note that clients depending on the local socket wait on xapi-wait-socket being
ready, which happens before the whole initialisation has finished.

Adjust the start timeout in the service to be the same as xapi-init-complete.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>