CA-371419: Always log exceptions when responding with 500 Internal Error
By default, logs from the http server are supressed, because it would
log far too much. However, it would be valuable to always log the
exception that has led to a 500 response, because this error condition
is often due to a bug.

Logging for these cases is done under a separate name, which is enabled
by default.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>