Do not log out session in xapi events loop to re-register VMs
The session.logout was used to interrupt the call to `event.from`, so
that is can be called again with a new list of VMs to watch, in case a
VM started, stopped etc.

The trouble is, that each such re-register event led to SESSION_INVALID
being logged, with stack trace, three times. This caused many people
looking at the logs unnecessary worries, while in reality the behaviour
is completely normal.

This replaces the register mechanism by adding the event loop's task
into the watch list for `event.from`, and injecting an event on the task
to interrupt the call and recompute the VM list.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>