Enable HTTPS for storage migration on the source
This is currently set on the destination, in host.migrate_receive.
However, all recent changes to make HTTPS migration work were on the
source host. Migration is allowed from older to newer software versions,
so it is possible that the destination is HTTPS capable, while the
source is not. This may result in the source receiving HTTPS URLs, which
is cannot handle, breaking this upgrade case.

Instead, let the source decide whether to switch to HTTPS or not,
depending on the config key (to be made the default later).

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>