CP-40357: Patch invalid lifecycles
These we all removed without being deprecated. These would be all
rejected by the lifecycle automaton.

Only with the SSL legacy it's clear at which released the message got
deprecated (as soon as it was introduced) For the rest I used an
easily-recognisable message that happens before and at the same release
as the removal.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>