CP-40357: Integrate state into the datamodel's lifecycle
This enables build the state at compile time without ad-hoc processing
as this is done in the datamodel itself. For example to block removed
fields from database constructors when autogenerating their code.

This information is encapsulated in Lifecycle.t which is a private
record. This makes it easy to deconstruct while forces its construction
using the function Lifecycle.from. This prevents inconsistent Lifecycles
from being created.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>