Push Storage_impl down to only wrap SMAPIv1 calls
This affects the entry point for SMAPIv2 calls in xapi, the server
module of which is currently defined in storage_mux.ml as

    module Server = Storage_interface.Server (Storage_impl.Wrapper (Mux)) ()

The Mux module does nothing other than forwarding the SMAPIv2 calls to
either a SMAPIv1 or a SMAPIv3 backend, depending on the SR type.

The wrapper in storage_impl.ml contains logic (including a state machine
and locking), which is mostly specific to SMAPIv1 backends. The current
SMAPIv3 backends also still rely on some of this, but the next
generation of SMAPIv3 will instead be incompatible with it.

This commit is the first step towards pushing the wrapper down to only
cover SMAPIv1 backends, the server module for which is defined in the
function `start_smapiv1_servers` in storage_access.ml. The file
storage_impl.ml is renamed to storage_smapiv1_wrapper.ml to reflect this
change. The Mux module now becomes the SMAPIv2 entry point.

There is a module called `Local_domain_socket` in storage_impl.ml, which
is unrelated to the wrapper. This is moved to storage_mux.ml.

There are a few things in the wrapper that are still needed for SMAPIv3,
even in the next generation. The following commits bring these back,
either in the Mux, or in the SMAPIv3 module (part of
xapi-storage-script).

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>