CP-33044 replace gpumon shutdown with NVML detach/attach
We want to avoid stopping gpumon while starting VMs with vGPUs because
it limits scalabiity and a daemon should run continiously. Use the
recently added capability to detach/attach the NVML library instead.

Introduce feature flag nvidia-gpumon-detach; when true, we use the new
attach/detach feature to keep gpumon running over the start of a VM.

This does not yet work with all GPUs but it would help cross-team
development to make this code available.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>