Xenctrl: drop interface_close
This got removed from Xen 4.18. We only use a global, so switch to using
Xenctrl.with_intf.

(We can't use a single global 'xc' value, because that might cause some unit
tests to fail since initializing xenctrl outside of a Xen system is not possible,
so delaying initialization is still needed).

Signed-off-by: Edwin Török <edvin.torok@citrix.com>