vhd-tool/direct_copy_stubs: fix setting of O_DIRECT flag
If a file descriptor is open in the normal way then its flags will be 0,
and the code here that attempts to enable O_DIRECT will do nothing.

This appears to have been an incorrect way of inverting an error condition
`ret < 0`, the negation of which is `ret >= 0`, and not `ret > 0`.

We open the file with O_DIRECT elsewhere though (in vhd-format), but fix this
code to avoid confusion.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>