CA-375276: xenctrlext_stubs.c: xc_get_last_error is not thread safe, use just errno which is
It is usually set to XC_INTERNAL_ERROR, except in xenguest, but we have no
bindings for that.

This API might get deleted in upstream Xen too: the error code is stored in the
xenctrl handle, but we made that global per process, and shared among all
threads. Which means that xenctrl calls from different threads can overwrite
each-others' error code.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>