Introduce DP.destroy2
This is a function that serves the same purpose as DP.destroy, but takes
additional SR, VDI and VM arguments besides the DP argument. This makes
it possible to implement it inside the SMAPIv3 backend, which otherwise
has no awareness of the concept of "datapaths" as used by xapi.

The Mux keeps a mapping from datapath to SR+VDI+VM. This is used to
transform DP.destroy calls into DP.destroy2 calls, which are sent to the
backend.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>