CA-373551: register for host events rather than task in events_from_xapi
A recent change registered for events on the task that was used by the
events_from_xapi thread in xapi_xenops. This is so that the event loop
can be woken up by injecting an event into this task, without the need
to logout its session (as previously done). The problem is that this
introduced a task in the DB that is forever "pending", which has
unwanted side-effects. For example, it gets GC'ed after 24 hours, and
also gets destroyed if the coordinator is restarted.

Switch to registering the host object for the localhost instead, which
should be more stable.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>