Add module DP_info to storage mux
This module maintains state for storage datapaths, in particular the SR,
VDI and VM their relate to, as was previously done in the wrapper in
storage_impl.ml, but in a much simplified way. This state is needed to
implement the functions in the DP module of SMAPIv2. The state is
persisted in a new file in /var/run/nonpersistent/xapi.

This is mainly to allow DP.destroy to be implemented.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>