Implement DP.destroy, VDI.attach and SR.reset in mux
A datapath is added to DP_info when the VDI is attached, and removed when
the VDI is detached or the datapath destroyed.

VDI.attach and attach2 are implemented in terms of attach3, as was done
in the old wrapper.

SR.reset just forwards to the backend.

Also some debug messages are copied from the wrapper, to improve
traceability.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>