CA-375271: Add static analyzer for OCaml C stubs
It looks for domain (runtime) lock usage, and shows an error whenever it
detects that an OCaml value is accessed, or an OCaml runtime function is called
without holding the lock.

It does not guarantee that the C stub is otherwise correct.

Also generates a primitives.h header that can be used to verify that the number
of arguments is correct for C stubs, and a primitives.model.c that generates
calls to all C stubs in pthreads, for the purposes of static analysis.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>