CP-31856: Option to use NBD to attach disks to the control domain
If the option `prefer-nbd-attach` is set to `true` in xapi.conf, then
use the `params_nbd` key that SM returns when attaching a VDI to connect
to the storage using NBD rather than the through the blktap2 kernel
device.

The VDI.attach call currently returns both `params` (for the blktap2
device) and `params_nbd` (for NBD).

The new mode is currently off by default, while we fix the remaining
functional problems. However, this will become the default in future,
such that we can drop blktap2 from our kernel.

Signed-off-by: Mark Syms <mark.syms@citrix.com>
Signed-off-by: Rob Hoes <rob.hoes@citrix.com>