CP-40650 Remove vtpm feature restriction
So far creation of vtpms is gated by the feature flag check. This needs
to be removed before it goes to general access so customers can have
access to the feature without friction.

Feature flags work like this:

* v6 provides a script to create a file: scripts/toggle-experimental-feature.sh
* this script creates an entry in /etc/xenserver/features.d
* this directory is read by v6
* the features/restrictions are read by xapi (from v6) and turned into
  an association list pool.restrictions
* this association list is checked by the code

The change keeps the core of the code but simply reports the state of
the feature flag. We might want to remove the code that sets it. It is
currently not set from within xapi.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>