CP-41796 Close Port 80 (Encrypt Data Transfer During VM Migrations)
This is a backport of several commits; we omit those that in the
original code deal with certificate checking as this is not present in
this code base.

The original code is in xen-api.git but in the backport is split into
xe-api.git, xenopsd.git, and xen-api-libs-transitional.git.

  5977845    Add HTTPS support to open-uri
  34e0b3b    Move Open_uri from xapi-idl to its own package
  d9f2d36    Move logging lib from xapi-idl to its own package
  140651b    Move good_ciphersuites from Xcp_consts to Constants
  0e92445    Rename Uuid module to Uuidx
  65f95bd    Do not wait for stunnel disconnect
  da27653    Use HTTP for local storage calls
  698f6e7    Use HTTPS when redirecting storage calls
  c30fc13    Switch host.migrate_receive over to HTTPS
  3910474    Switch VM.pool_migrate over to HTTPS

Becasue the actual change is not large, I am keeping all changes in a
single commit.

The original code was authored by Rob Hoes <rob.hoes@citrix.com>. Almost
all code had to be ported manually.

This switches the xenopsd-to-xenopsd connection over to HTTPS, if
enabled in the config file (currently off by default).

Socket keepalives do not work when stunnel is used, as the given fd is
the local connection to stunnel, and are not set in HTTPS mode. This is
fine, because the stunnel client is already set up with keepalives.

The host.migrate_receive call returns URLs for the sending host to use
in its VM.migrate_send call. HTTPS URLs are returned based on the value
of the config option.

The function in question redirects a local storage call to another host
if the current host does not have access to the SR. This involves
rewriting a localhost HTTP URL to a remote URL, which must be an HTTPS
URL if storage migration is configured to use HTTPS.

Storage calls that are handled by the local host should always use HTTP,
and not set up a connecting through stunnel. The SM URL that is passed
to VM.migrate_send in the `dest` parameter, which comes from the return
value of a call to host.migrate_receive, may be an HTTPS URL to the
localhost, which needs to be rewritten upon receiving it.

When starting a mirroring process for a disk as part of a storage
migration, xapi establishes the connection to the destination, but hands
it over to tapdisk to do the actual mirroring over NBD. It is crucial
that xapi just hands over the file descriptor and then continues with
other business, without waiting for the connection to finish. This is
how it works for TCP connections now.

When switching to TLS connections, xapi starts an stunnel process as
part of the connection setup, and hands over the stunnel fd to tapdisk.
By default, this functionality then waits for stunnel to finish after
the connection is eventually broken, thus introducing the unwanted
blocking. We fix this by telling stunnel to disconnect, but not wait for
this to actually happen.

This is currently set on the destination, in host.migrate_receive.
However, all recent changes to make HTTPS migration work were on the
source host. Migration is allowed from older to newer software versions,
so it is possible that the destination is HTTPS capable, while the
source is not. This may result in the source receiving HTTPS URLs, which
is cannot handle, breaking this upgrade case.

Instead, let the source decide whether to switch to HTTPS or not,
depending on the config key (to be made the default later).

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>
Signed-off-by: Christian Lindig <christian.lindig@cloud.com>