python/XenAPI: Replace import six.moves with stdlib imports
Replace the use of six.moves with directly importing the
standard xmlrpc and http libs of Python 2.7 and Python 3.x

Tested by using xenserver-status-report/xen-bugtool with it
using on XS 8.3.0 with Python 2.7 and Python 3.6 installed.

Consequently, the Requires: python-six and python3-six can
be removed from the xen-api.spec file as well.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@citrix.com>