CA-375256: Fix storage initialisation on xapi startup
The top-level SMAPIv2 server in xapi uses a mux to forward calls to
lower level SMAPI servers depending on the SR. This relies on having
each SR registered with the mux as soon as the related PBD is plugged on
a host. These registrations are held in memory and therefore need to be
refreshed every time xapi starts.

Previously, the mux had a wrapper on top of it, which handled some of
the SMAPIv2 calls itself, without forwarding invoking the mux at all. An
example is the SR.list call, which returns the list of currently
attached SRs on a host. This used state from a file in dom0 and
therefore worked before registering SRs with the mux.

Unfortunately, `SR.list` is used a few times in xapi's startup sequence
before mux registrations are complete. Now that the wrapper has been
removed, the function always returns an empty list until the mux is
fully set up. This breaks several bits of initialisation at dbsync time,
in particular in the `resynchronise_pbds` function in storage_access.ml.

This fix ensures that all SRs are registered with the mux before calling
functions on the top-level SMAPIv2 server.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>