CP-41675: fix idempotent behaviour of Helpers.FileSys.rmrf
If a file has already been deleted using rmrf, using it again
currently raises a ENOENT exception. This is not the expected
behaviour of rm -rf.

This patch ensures that for rmrf, ENOENT is a noop as expected.

The few modules in xapi using rmrf expect no file after calling rmrf,
so this change does not change their expectation for the rmrf behaviour.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>