CP-42007: platform:secureboot=auto means platform:secureboot=true always
When the uefi certificates are present in the pool.uefi-certificates field
of the xapi db:
1) During vm-start, platform:secureboot=auto means platform:secureboot=true
2) During vm-create in XenCenter, the UEFI Secure Boot checkbox is enabled
3) other clients may rely on this field being populated to decide if/how
they should set the VM platform:secureboot field.

To maintain compatibility with this existing behaviour of these components,
in case override_uefi_certs=false (which is default in xenserver), then
during xapi-start do populate pool.uefi-certificates with the uefi
certificates from /var/lib/varstored, which links to the uefi certificates
statically installed by RPMs.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>