CP-40388: store attach mode (rw/ro) with datapath in mux
We do this, because in SMAPIv2, the rw/ro mode is part of the VDI attach
call, while in SMAPIv3 it belongs to VDI activate. The activate call in
SMAPIv2 does not have a mode parameter. Storing the mode with the
datapath when attaching allows to switch case when handling the
following activate call.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>