CP-40388: Add VDI.activate_readonly to the storage interface
This is a new call in the SMAPIv3. The easiest way to use it is to add a
similar call to the SMAPIv2 and switch case in the mux. This is done
only if the SR backend has advertised the new VDI_ACTIVATE_READONLY
feature.

The SMAPIv1 has the rw/ro mode associated with VDI.attach rather than
activate, so activate_readonly simply calls activate in the SMAPIv1
backend.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>