python: Use xapi's versioning scheme for XenAPI package
This allows us to do releases much more frequently as pipy's policy
is to have immutable packages.

It also simplifies the build process quite a bit, which is neat.

The current method of building allows to build the package under git and
when git information is not available, using the same mechanism as xapi.
(using XAPI_VERSION if it's set, with a fallback on git, or failing
miserably)

This is not enough to be able to buuld it in a .spec environmtent
because the version schemas generated for the spec file conflict with
the allowed versions of python packages.

e.g.
koji generates
  23.5.0.19.g3657e16cb
python rejects it, it would accept
  23.5.0.19+g3657e16cb
the autogenerated equivalent generated by python would be
  23.5.0.dev19+g3657e16cb

This changeset is not meant to resolve this conflict, as the current
solution is good enough for its main use-case: generating a release
using github actions

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>