CP-39935 implement VTPM export
A VM is exported as a series of XML objects each of which is created
using the RPC mechanism. The VTPM data is actually split into two
objects: VTPM and content; we export them together in one record. To do
so, we need to define a custom type and its corresponding RPC
representation and make sure a VM's VTPM gets exported.

The records in the XML representation can refer to each other. So part
of the export a table is constructed that provides a mapping between
object references as used in the database and export-local fresh ones.
We need to enter the VTPM record into this table such that the VM record
can refer to it.

For now we don't pay attention to the persistence backend on import.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>