CP-39935 Implement VTPM import
On import, always create a new VTPM object for a VM and don't look for
an existing object.

We are importing a custom record type vtpm' that contains the content of
the VTPM; hence, we use vtpm'_of_rpc to read it.

We are not paying attention to the backend - we import the data and
possibly store it in a different backend from the one which exported
this VTPM. Hence, we anticipate changes of the backend in the future.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>