CP-39863 add VTPM "allowed operations" support
Add field "allwoed operations" to VTPM objects and keep it updated. This
is for API clients to avoid calling method that would fail in the
current state. For the VTPM object allowed operations only depend on the
state of its VM. Update allowed operations when the VM update its
allowed operations.

When the datamodel is updated we must update its hash value. To run this
check (which is part of the wider unit tests) in isolation, use "make
schema".

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>