xapi_metastore: add basic configuration file with deserialization tests
Use a `V1 of ...` variant internally, so we can change the type and keep
the ability to read old configuration files.

Automatically generate some deserialization tests (like we do for RPCs, except
for records).

TODO: need more quicktests, also check that etcd is healthy, etc.

There are a lot more config option, they will be part of separate commits,
these are the minimum needed to start etcd on a single host.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>