[maintenance]: repeat test errors at the end
While developing locally it is useful to see the errors as early as possible
(which is the default).
However in the CI the error will be mixed together with all the output from the
successful tests and locating the error message needs some scrolling.
Dune has a mode where it does both: report the error as soon as it is
discovered, but also repeat just the errors again at the end, which will make
it more obvious in the CI on what the error is.

For this to be useful turn output buffering back on (otherwise dune will only
report the failing testcase at the end but you still have to scroll back to
find the error mixed with the output from all other tests).

Signed-off-by: Edwin Török <edwin.torok@cloud.com>