CA-375358: Parse output of yum upgrade to get RPMs to be updated/installed
In getting which RPM packages are going to be updated and installed
before applying updates on a host, following cases are considered:
1. package-A-v1-r1 is installed and package-A-v2-r1 is an update. A new
   package-B-v1-r1 is required (hence introduced) by package-A-v2-r1.
2. package-X-v1-r1 is required by package-Y-v1-r1, and both are
   installed. Now package-Y-v2-r1 is an update, and doesn't require
   package-X anymore.

In case#1, package-B-v1-r1 can't be returned by "yum list updates", but
it can be returned by "yum list available". This was fixed in commit
c020034a21e. The fix results in another issue that in case#2, the
package-X-v1-r1 is returned as an update from "yum list available".
Ideally the package-X-v1-r1 should be obsoleted so that the fix in
commit c020034a21e can handle it correctly. But it would be better to
make the XAPI get the packages to be updated and installed exactly as
what will be done in applying updates by "yum upgrade".

This approach was considered in the first place when this feature was
implemented. But the difficulty is that the output of "yum upgrade"
can't be formatted as other YUM commands.
To avoid more potential issues in future, this commit adds function to
parse the RPM packages to be updated and installed from output of
dry-run of "yum upgrade".

Signed-off-by: Ming Lu <ming.lu@cloud.com>