CA-376294: Extract hostname from FQDN
The local hostname may be set to FQDN, which is likely to exceed
the netbios limit (15 chars), Then the netbios is compressed and
randomized to meet the limit.
However, user may pre-create the machine account to join domain
with the hostname part of FQDN, thus we need to extract the
hostname from the FQDN as netbios to join domain.

Signed-off-by: Lin Liu <lin.liu@citrix.com>