CA-376319: Ensure that nbd_client_manager cannot block forever.
This has been seen to occur in the prescence of a kernel bug which
resulted in the NBD device state not being correctly updated. The fix
for which is obviously to update the kernel to not have the bug, but
also make the calling code robust to the condition and not block for
an infinite amount of time.

Remove the lock surrounding the disconnect operation as it is are not
required and provides no benefit. Locks are only required to prevent
contention between multiple clients trying to acquire the same NBD
device having determined that it currently unused.

Signed-off-by: Mark Syms <mark.syms@citrix.com>