CA-375358: Add redundancy in getting latest updates/installations
This commit adds a redundancy for getting latest updates/installations
from both 'repoquery' and 'yum upgrade (dry run)'. This is because:
1. the outputs of 'repoquery' are formatted which is good, but they
   don't return the updates/installations exactly;
2. the output of 'yum upgrade (dry run)' can return the
   updates/installations exactly, but it is non-formatted which would
   result in fragile parsing logic.

Since this is critical to update function, an error in the fragile
parsing might prevent the system from being applied furhter bugfixes
and consequently make the update function stranded completely.
The redundancy is to ensure that when any error occurred in the fragile
parsing, the function would fall back to use the outputs of 'repoquery'.

Signed-off-by: Ming Lu <ming.lu@cloud.com>