CP-42835: Allow changing DNS servers when HA or clustering enabled
If HA or clustering is enabled, we are at risk of fencing if the IP
address, netmask or gateway changes. This may happen if we switch to
DHCP mode, or if we change a static configuration.

The current checks deny any IP configuration change. However, changing
just DNS servers in static mode would be safe even if HA or clustering
is enabled. This commit relaxes the check.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>