ci: reduce code run with permissions to release to PyPI
Now only the official PyPI action is run with permissions to upload.

Unfortunately to share the files between the job that builds it and the two
jobs that use them, some bureaucracy is needed in the form of "artifacts".
These need to be uploaded when produced and downloaded when read. On top of
that the dependencies needs to be manually defined instead of just inferred, so
just do that. Use that to wait to publish to PyPI until the release in github
has been done.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>