CP-40528 VTPM snapshot, revert, clone
Implement VM snapshot/revert and VM clone for VMs that have VTPMs.

A VM snapshot is a VM record with an additional field snapshot_metadata
that contains all metadata of the VM we are taking a snapshot of. This
metadata includes a reference to the VTPM the VM is/was using. In
addition, a new VTPM as a copy of the existing one is created and
referred to by the VM record that represents the snapshot. So a snapshot
references two VTPM: one directly (A) and one via snapshot_metadata (B).

Since the VM that took the snapshot is in continued use, it will
continue to update B but A is stable.

When reverting the snapshot we create a new VM from the VM record that
represents the snapshot. This can happen again and again.  Unlike a
clone, reverting a snapshot does not consume it.

As a consequence, the VM we are creating can't use VTPM A and modify it
but we create a copy of A and use that.  The copy (of the VTPM) has a
UUID that is different from the UUID before the revert. This is
deliberate because maintaining the UUID is not done in other cases
either. In general, the snapshot and the current VM can a have a
different number of VTPMs.

To help development, this patch extends the fields listed in "xe
vtpm-list" by showing the size and MD5 hash of the content. Otherwise
tracking this is difficult because the content is not stored in the VTPM
object but it refers to another object of class secret that holds the
actual content.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>