xapi-stdext-encodings: drop encoding and unused functions from API
XAPI only uses this module for validating XML UTF8 strings.
However all the unit tests here are still useful, so gradually drop
functionality from the encodings API until we're left just with the
validator, which can be implemented more efficiently without allocation.

This moves some code that is still useful for tests into the test
module (which can be used to check that the stdlib or Uutf's semantics
matches the pre-existing semantics of this module).

Signed-off-by: Edwin Török <edwin.torok@cloud.com>