xapi-stdext-encodings: drop all functions that allocate from the interface
Moved decoding into the test code, and supply only an UCS validator as
input to the functor.
This will allow replacing the implementation with a more efficient one
based on Uchar.t and Uutf (or the stdlib in 4.14)

Drops some tests that no longer makes sense, but still keep the majority
of the tests.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>